api/requirements.txt
# Azure Functions
azure-functions

# Web scraping
requests
beautifulsoup4

# Azure Table Storage
azure-data-tables
tzdata